Separate argsutils functions
